BlogByLLAMA/requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
langchain-community
python-box
streamlit
python-dotenv